+++ routes/routes.js
@@ -1,13 +1,15 @@
 const router = require("express").Router();
 const apiRoute = require("./subroutes/apiRoute");
-const overviewRoute = require("./subroutes/overviewRoute");
+const homeRoute = require("./subroutes/homeRoute");
 const adminRoute = require("./subroutes/adminRoute");
+// const categoryRoute = require("../subroutes/categoryRoute");
 
 router.get("/about", function (req, res) {
   res.render("pages/about");
 });
 
-router.use("/home", overviewRoute);
+router.use("/home", homeRoute);
+// router.use("/overview", overviewRoute);
 router.use("/admin", adminRoute);
 router.use("/api", apiRoute);
