+++ routes/subroutes/homeRoute.js
@@ -1,6 +1,7 @@
 const router = require("express").Router();
-const overviewController = require('../../controllers/overviewController')
+const overviewController = require('../../controllers/homeController')
 
 router.get("/", overviewController.getIndex);
+router.get("/")
 
 module.exports = router;
